requirements.txt
beautifulsoup4
lxml
requests
pillow
flask
celery
sqlalchemy

green
nose
requests-cache
